pyproject.toml
[tool.poetry]
name = "gan-htr"
version = "0.1.0"
description = "GAN-HTR: A Generative Adversarial Network for Handwritten Text Recognition"
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10, <3.12"
tensorflow = "2.16.1"
tensorflow-datasets = "4.9.4"
tensorflow-text = "2.16.1"
tensorflow-io = "0.37.0"
tensorflow-gan = "2.1.0"
dm-sonnet = "2.0.2"
editdistance = "0.8.1"
matplotlib = "3.9.0"
opencv-python = "4.9.0.80"
pillow = "10.3.0"
tqdm = "4.66.4"
scipy = "1.13.1"
numpy = "1.26.4"
h5py = "3.11.0"
aiohttp = "3.9.5"
astor = "0.8.1"
async-timeout = "4.0.3"
attrs = "23.2.0"
blinker = "1.8.2"
brotlipy = "0.7.0"
cachetools = "5.3.3"
certifi = "2024.7.4"
cffi = "1.16.0"
chardet = "5.2.0"
click = "8.1.7"
cryptography = "42.0.8"
cycler = "0.12.1"
gast = "0.5.4"
google-auth = "2.30.0"
google-auth-oauthlib = "1.2.0"
google-pasta = "0.2.0"
grpcio = "1.64.1"
idna = "3.7"
importlib-metadata = "8.0.0"
keras = "3.3.3"
kiwisolver = "1.4.5"
markdown = "3.6"
mkl-fft = "1.3.8"
mkl-random = "1.2.4"
mkl-service = "2.4.0"
multidict = "6.0.5"
oauthlib = "3.2.2"
opt-einsum = "3.3.0"
protobuf = "4.25.3"
pyasn1 = "0.6.0"
pyasn1-modules = "0.4.0"
pycparser = "2.22"
pyjwt = "2.8.0"
pyopenssl = "24.1.0"
pyparsing = "3.1.2"
pysocks = "1.7.1"
python-dateutil = "2.9.0"
requests = "2.32.3"
requests-oauthlib = "2.0.0"
rsa = "4.9"
six = "1.16.0"
tensorboard = "2.16.2"
tensorboard-plugin-wit = "1.8.1"

termcolor = "2.4.0"
urllib3 = "2.2.1"
werkzeug = "3.0.3"
wrapt = "1.16.0"
yarl = "1.9.4"
zipp = "3.19.2"
kagglehub = {extras = ["hf-datasets"], version = "^0.3.12"}
datasets = "^4.0.0"
imageio = "^2.37.0"
augraphy = {extras = ["all"], version = "^8.2.6"}

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
